feat: MainPageClient 컴포넌트 추가 및 페이지 구조 리팩토링
