+++ src/app/main/_section/mainHeader/Carousel/mainCarousel.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import useEmblaCarousel from "embla-carousel-react";
 import { useEffect, useState } from "react";
 import * as styles from "./mainCarousel.css.ts";
@@ -6,6 +8,8 @@ import Image, { StaticImageData } from "next/image";
 interface MainCarouselProps {
   images: { id: string; src: StaticImageData | string }[];
 }
+
+const CAROUSEL_SIZES = "(max-width: 76.8rem) 100vw, 76.8rem";
 
 const MainCarousel = ({ images }: MainCarouselProps) => {
   const [emblaRef, emblaApi] = useEmblaCarousel({ loop: true });
@@ -38,9 +42,20 @@ const MainCarousel = ({ images }: MainCarouselProps) => {
   return (
     <div className={styles.carouselContainer} ref={emblaRef}>
       <div className={styles.carouselSlides}>
-        {images.map((image) => (
+        {images.map((image, index) => (
           <div key={image.id} className={styles.carouselSlide}>
-            <Image src={image.src} alt={`Slide ${image.id}`} className={styles.image} width={390} height={156} />
+            <Image
+              src={image.src}
+              alt={`Slide ${image.id}`}
+              className={styles.image}
+              width={390}
+              height={156}
+              sizes={CAROUSEL_SIZES}
+              // priority: <link rel="preload"> 추가
+              // fetchPriority: img + preload 요청 모두에 high 적용 (Next 15부터 분리됨)
+              priority={index === 0}
+              fetchPriority={index === 0 ? "high" : "auto"}
+            />
           </div>
         ))}
       </div>
